RandomizeGMService -> GMGenerationService refactoring
